Forget current profile on match_and_apply failure
When an output changes and match_and_apply() attempts to match a new
profile, any failure to match (or apply) a new profile would prevent
re-application of any previously matched profile, even if it matches
after a subsequent output change. The right conceptual behavior is to
forget a no-longer-applicable profile after match_and_apply fails, which
fixes this issue.

Fixes: https://todo.sr.ht/~emersion/kanshi/100