parser: implements global output defaults
This commit allows the output directive to appear in the global scope.

Using the output directive in the global scope sets defaults for that
output that apply to every profile the output is mentioned in.

Output defaults are resolved in a second step after the config is parsed
to keep the config syntax declarative.