parser: implements output alias directive
This commit allows output default directives in global scope to set
aliases for outputs.

Output aliases must start with a '$' sign to be able to raise an error when
output aliases are mistyped. Output aliases can only be declared in
global scope and can only be used in profile scope.

Closes: https://todo.sr.ht/~emersion/kanshi/106